MARTIFACT-17: Add generics details to collections Submitted by: Jason Dillon
git-svn-id: https://svn.apache.org/repos/asf/maven/artifact/trunk@658725 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/maven/artifact/versioning/ComparableVersion.java
@@ -77,14 +77,14 @@ public class ComparableVersion
 
         public boolean isNull()
         {
-            return ( value.intValue() == 0 );
+            return (value == 0 );
         }
 
         public int compareTo( Item item )
         {
             if ( item == null )
             {
-                return value.intValue() == 0 ? 0 : 1; // 1.0 == 1, 1.1 > 1
+                return value == 0 ? 0 : 1; // 1.0 == 1, 1.1 > 1
             }
 
             switch ( item.getType() )
